Added Makefile setup to compile libft and printf proe

+++ ft_printf/ft_puthex_fd.c
@@ -6,14 +6,13 @@
 /*   By: pamallet <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 17:00:11 by pamallet          #+#    #+#             */
-/*   Updated: 2024/11/06 18:58:51 by pamallet         ###   ########.fr       */
+/*   Updated: 2024/11/07 00:18:11 by paul_mall        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
-#include "libft/libft.h"
 
-void	ft_puthex_fd(unsigned int n, char *shex, int fd)
+void	ft_puthex_fd(long n, char *shex, int fd)
 {
 	if (n >= 0 && n <= 15)
 		write(fd, &shex[n % 16], 1);
